% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimEst.R
\name{.opModel}
\alias{.opModel}
\title{Surplus production operating model}
\usage{
.opModel(obj, seed = 1)
}
\description{
\code{.opModel} is a hidden function that generates a biomass
time series from the parameters of a control object
}
\details{
Details details
}
