% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runSimEst}
\alias{runSimEst}
\title{Run the simulation-estimation procedure}
\usage{
runSimEst(ctlFile = "simCtlFile.txt", folder = NULL, quiet = TRUE)
}
\arguments{
\item{ctlFile}{Character. Path to sim control file.}

\item{folder}{Character. Prefix to append to output folder name}

\item{quiet}{Logical. Suppress hierProd messages?}
}
\value{
NULL. 

   Creates sim folder with results in an RData file, 
   a copy of the msProd.cpp TMB template, and a copy of the control
   file.
}
\description{
\code{runSimEst} parses a control file and fits hierProd
to simulated data.
}
\details{
This is the workhorse function, and can be called directly from
the console for running single cases, and is also used by
batch functions when running full factorial experiments
}
\examples{
# Run a simulation with the base control file
runSimEst()

# Run a simulation with the base control file,
# but include a custom directory for output
rimSimEst( folder = "testDir" )

\dontrun{
 runSimEst(ctlFile = NULL)
}
}
